Foolproof Design
Implemented foolproof design for the undo and redo buttons, and the move up and down buttons. No changes were needed for the add item, close list, and trash list buttons because they are already fool proof. The same goes for the delete item button because it only exists when there is an item to delete.