Assigned To Field and New Column Spacing
Implemented the assigned to field. When a new item is created the default value is "Not Assigned". In order for the assigned to field to work it was necessary to implement the new column spacing as well. The description field now has a value of 3 while the other column spaces all have values of 2. While adding the assigned to field, I have also implemented function undo/redo for the editing of this field.